move xxxGetHash to hash.c
