+++ IDE/IAR-EWARM/Templates/wolfSSL-Lib/wolfSSL-Lib.eww
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="iso-8859-1"?>
+
+<workspace>
+  <project>
+    <path>$WS_DIR$\wolfSSL-Lib.ewp</path>
+  </project>
+  <batchBuild/>
+</workspace>
+
+
